rework logic for EXT-X-PROGRAM-DATE-TIME insertion in hls playlists
To allow multiples media_playlists to be synchronized a reference time
is generated in simple_hls_notifier.

Reference time is computed from local time + an offset specified in the
hls_ext_x_program_date_time option.

Each time a segment is added in MediaPlaylist, we now pass the
reference_time, so that segment can compute the 'wall time' associated
with this segment.

Reference time is reset on each discontinuity

In playlist, EXT-X-PROGRAM-DATE-TIME is only inserted for first segment,
and segments following a discontinuity.